Factor out core date widget functionality
